requirements.txt
# Core
lyzr-adk>=0.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Storage
aiosqlite>=0.20.0

# MCP (for diagram tool — testing Lyzr + MCP bridge)
mcp>=1.0.0

# Diagram generation (fallback when MCP server unavailable)
diagrams>=0.23.4

# HTTP client (for lyzr_docs_tool)
httpx>=0.27.0

# Utilities
python-multipart>=0.0.9
aiofiles>=24.1.0
